Refactors and extracts Country class
