+++ app/public/css/mainstyle.css
@@ -62,6 +62,8 @@ body {
   text-align: center;
   float:left;
   font-family: fantasy;
+  background-color:rgba(0, 0, 0, 0.2);
+  border-radius:15%;
 
 }
