+++ app/views/index.erb
@@ -47,7 +47,7 @@
           <input type='radio' name='no_of_players' value='two_player'> Two Players <br>
         </div>
         <div class='button-container'>
-          <input class='start-button' type='submit' value='PLAY'> <br>
+          <input class='start-button' type='submit' value='READY'> <br>
           <div class='player-notice'>
             <p class='mainpage-notice'><%= @actors[0] %> doesn't know how to play Rock, Paper, Scissors,<br>
             Lizard, Spock and needs your help to win the election.</p>
