+++ app/views/play.erb
@@ -8,7 +8,7 @@
     </style>" %>
   </head>
   <body>
-    <%= "<audio src='audio/vocals/#{@game.winner}_wins.wav' autoplay='autoplay'></audio>" %>
+    <%= "<audio src='audio/vocals/#{@game.winner_name}_wins.wav' autoplay='autoplay'></audio>" %>
     <div class='page-container'>
       <h1 class='heading'>ROCK, PAPER, POLITICS!</h1>
       <div class='banner-container'>
