+++ spec/features/web_helpers.rb
@@ -4,7 +4,7 @@ def sign_in_and_play
     fill_in 'player', with: :Trump
     find("input[value='Rock']").click
   end
-  click_button 'PLAY'
+  click_button 'READY'
 end
 
 def sign_in_two_players_and_play
@@ -14,5 +14,5 @@ def sign_in_two_players_and_play
     find("input[value='Rock']").click
     find("input[value='two_player']").click
   end
-  click_button 'PLAY'
+  click_button 'READY'
 end
